Separate previous and next plugins
