+++ demo/latest/webpack.config.js
@@ -54,7 +54,8 @@ module.exports = {
             '@phuocng/rpv': path.join(__dirname, '../../packages/rpv/src'),
             '@phuocng/rpv-current-page': path.join(__dirname, '../../packages/rpv-current-page/src'),
             '@phuocng/rpv-locale-switcher': path.join(__dirname, '../../packages/rpv-locale-switcher/src'),
-            '@phuocng/rpv-page-navigation': path.join(__dirname, '../../packages/rpv-page-navigation/src'),
+            '@phuocng/rpv-next-page': path.join(__dirname, '../../packages/rpv-next-page/src'),
+            '@phuocng/rpv-previous-page': path.join(__dirname, '../../packages/rpv-previous-page/src'),
         },
     },
     devServer: {
