+++ packages/rpv-next-page/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@phuocng/rpv-page-navigation",
+    "name": "@phuocng/rpv-next-page",
     "version": "2.0.0",
     "description": "A React component to view a PDF document",
     "license": "https://react-pdf-viewer.dev/license",
@@ -19,7 +19,7 @@
     "repository": {
         "type": "git",
         "url": "https://github.com/phuoc-ng/react-pdf-viewer",
-        "directory": "packages/rpv-page-navigation"
+        "directory": "packages/rpv-next-page"
     },
     "bugs": {
         "url": "https://github.com/phuoc-ng/react-pdf-viewer/issues"
