--- packages/rpv-page-navigation/src/StoreProps.ts
@@ -1,7 +1,0 @@
-interface StoreProps {  
-    currentPage?: number;
-    jumpToPage?(pageIndex: number): void;
-    numberOfPages?: number;
-}
-
-export default StoreProps;
