+++ packages/rpv-previous-page/package.json
@@ -1,0 +1,48 @@
+{
+    "name": "@phuocng/rpv-previous-page",
+    "version": "2.0.0",
+    "description": "A React component to view a PDF document",
+    "license": "https://react-pdf-viewer.dev/license",
+    "author": {
+        "name": "Nguyen Huu Phuoc",
+        "email": "me@phuoc.ng",
+        "url": "https://twitter.com/nghuuphuoc"
+    },
+    "homepage": "https://react-pdf-viewer.dev",
+    "keywords": [
+        "react",
+        "react.js",
+        "pdf",
+        "pdf.js",
+        "pdf viewer"
+    ],
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/phuoc-ng/react-pdf-viewer",
+        "directory": "packages/rpv-previous-page"
+    },
+    "bugs": {
+        "url": "https://github.com/phuoc-ng/react-pdf-viewer/issues"
+    },
+    "dependencies": {
+        "react": "^16.13.1",
+        "react-dom": "^16.13.1"
+    },
+    "devDependencies": {
+        "@rollup/plugin-json": "^4.1.0",
+        "@rollup/plugin-typescript": "^5.0.0",
+        "@types/react": "^16.9.41",
+        "@types/react-dom": "^16.9.8",
+        "@typescript-eslint/eslint-plugin": "^3.4.0",
+        "@typescript-eslint/parser": "^3.4.0",
+        "cssnano": "^4.1.10",
+        "eslint": "^7.3.1",
+        "eslint-plugin-react": "^7.20.0",
+        "postcss-cli": "^7.1.1",
+        "rollup": "^2.18.0",
+        "rollup-plugin-less": "^1.1.2",
+        "rollup-plugin-less-modules": "^0.1.11",
+        "rollup-plugin-terser": "^6.1.0",
+        "typescript": "^3.9.5"
+    }
+}
