+++ packages/rpv-previous-page/src/PreviousPageButton.tsx
@@ -1,62 +1,64 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2020 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
 import React, { useContext, useEffect, useState } from 'react';
-import { Button, LocalizationContext, NextIcon, Position, Store, StoreHandler, Tooltip } from '@phuocng/rpv';
+import { Button, LocalizationContext, Position, PreviousIcon, Store, StoreHandler, Tooltip } from '@phuocng/rpv';
 
 import StoreProps from './StoreProps';
 
-export interface RenderNextPageButtonProps {
+export interface RenderPreviousPageButtonProps {
     isDisabled: boolean;
     onClick: () => void;
 }
 
-export interface NextPageButtonProps {
-    children?: ChildrenNextPageButton;
+export interface PreviousPageButtonProps {
+    children?: ChildrenPreviousPageButton;
 }
 
-export type ChildrenNextPageButton = (props: RenderNextPageButtonProps) => React.ReactElement;
+export type ChildrenPreviousPageButton = (props: RenderPreviousPageButtonProps) => React.ReactElement;
 
 const TOOLTIP_OFFSET = { left: 0, top: 8 };
 
-const NextPageButton: React.FC<{
-    children?: ChildrenNextPageButton,
+const PreviousPageButton: React.FC<{
+    children?: ChildrenPreviousPageButton,
     store: Store<StoreProps>,
 }> = ({ store, children }) => {
     const l10nContext = useContext(LocalizationContext);
-    const [numberOfPages, setNumberOfPages] = useState(0);
     const [currentPage, setCurrentPage] = useState(0);
 
-    const handleNumberOfPages: StoreHandler<number> = (n: number) => {
-        setNumberOfPages(n);
-    };
     const handleCurrentPageChanged: StoreHandler<number> = (currentPage: number) => {
         setCurrentPage(currentPage);
     };
 
     useEffect(() => {
         store.subscribe('currentPage', handleCurrentPageChanged);
-        store.subscribe('numberOfPages', handleNumberOfPages);
 
         return () => {
             store.unsubscribe('currentPage', handleCurrentPageChanged);
-            store.unsubscribe('numberOfPages', handleNumberOfPages);
         };
     }, []);
 
-    const goToNextPage = () => {
+    const goToPreviousPage = () => {
         const jumpToPage = store.get('jumpToPage');
         if (jumpToPage) {
-            jumpToPage(currentPage + 1);
+            jumpToPage(currentPage - 1);
         }
     };
 
-    const defaultChildren = (props: RenderNextPageButtonProps) => {
+    const defaultChildren = (props: RenderPreviousPageButtonProps) => {
         const label = (l10nContext && l10nContext.plugins)
-            ? l10nContext.plugins.pageNavigation.nextPage
-            : 'Next page';
+            ? l10nContext.plugins.pageNavigation.previousPage
+            : 'Previous page';
 
         return (
             <Tooltip
                 position={Position.BottomCenter}
-                target={<Button onClick={props.onClick}><NextIcon /></Button>}
+                target={<Button onClick={props.onClick}><PreviousIcon /></Button>}
                 content={() => label}
                 offset={TOOLTIP_OFFSET}
             />
@@ -65,9 +67,9 @@ const NextPageButton: React.FC<{
     const render = children || defaultChildren;
 
     return render({
-        isDisabled: currentPage + 1 >= numberOfPages,
-        onClick: goToNextPage,
+        isDisabled: currentPage <= 0,
+        onClick: goToPreviousPage,
     });
 };
 
-export default NextPageButton;
+export default PreviousPageButton;
