+++ packages/rpv-previous-page/src/StoreProps.ts
@@ -1,0 +1,14 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2020 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+interface StoreProps {
+    currentPage?: number;
+    jumpToPage?(pageIndex: number): void;
+}
+
+export default StoreProps;
